MessageFormat: Use off_t argument instead of int32
Which doubles the maximum width fixing #11809. I wanted to convert this to
a variadic functionbut that would require updating all of the instance that use
MessageFormat or a clever hack to get around the ambiguity of an off_t
argument and an int argument count with 0 arguments. Either way seemed
out of the scope of fixing this bug.

Fixes #11809

Another Beta1 bug down!

+++ headers/os/locale/MessageFormat.h
@@ -1,6 +1,10 @@
 /*
- * Copyright 2014, Haiku, Inc.
+ * Copyright 2014-2015 Haiku, Inc. All rights reserved.
  * Distributed under the terms of the MIT License.
+ *
+ * Authors:
+ *		Adrien Destugues, pulkomandy@pulkomandy.tk
+ *		John Scipione, jscipione@gmail.com
  */
 #ifndef _B_MESSAGE_FORMAT_H_
 #define _B_MESSAGE_FORMAT_H_
@@ -24,7 +28,7 @@ public:
 
 			status_t			InitCheck();
 
-			status_t			Format(BString& buffer, const int32 arg) const;
+			status_t			Format(BString& buffer, const off_t arg) const;
 
 private:
 			status_t			_Initialize(const icu::UnicodeString&);
